Adds kmeans and weighted-centroid
