+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-form-models/base-analysis-form-model.js
@@ -100,6 +100,29 @@ module.exports = Backbone.Model.extend({
   _setSchema: function (schema) {
     this.schema = schema;
     this.trigger('changeSchema', this);
-  }
+  },
 
+  _getColumnsByType: function (columnType) {
+    var nodeDefModel = this._layerDefinitionModel.findAnalysisDefinitionNodeModel(this.get('source'));
+    var querySchemaModel = nodeDefModel.querySchemaModel;
+
+    var sourceColumns = querySchemaModel.columnsCollection.map(function (columnModel) {
+      var columnName = columnModel.get('name');
+      var column = {
+        val: columnName,
+        label: columnName,
+        type: columnModel.get('type')
+      };
+
+      return column;
+    });
+
+    if (columnType) {
+      sourceColumns = sourceColumns.filter(function (column) {
+        return column.type === columnType;
+      });
+    }
+
+    return sourceColumns;
+  }
 });
