+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-form-models/weighted-centroid-form-model.js
@@ -1,3 +1,4 @@
+var _ = require('underscore');
 var BaseAnalysisFormModel = require('./base-analysis-form-model');
 var template = require('./weighted-centroid-form.tpl');
 
@@ -8,6 +9,7 @@ var template = require('./weighted-centroid-form.tpl');
 module.exports = BaseAnalysisFormModel.extend({
   initialize: function () {
     BaseAnalysisFormModel.prototype.initialize.apply(this, arguments);
+    this.on('change:aggregate', this._setSchema, this);
     this._setSchema();
   },
 
@@ -20,7 +22,7 @@ module.exports = BaseAnalysisFormModel.extend({
    * @override {BaseAnalysisFormModel._setSchema}
    */
   _setSchema: function () {
-    BaseAnalysisFormModel.prototype._setSchema.call(this, {
+    var schema = {
       source: {
         type: 'Select',
         text: _t('editor.layers.analysis-form.source'),
@@ -30,35 +32,34 @@ module.exports = BaseAnalysisFormModel.extend({
       weight_column: {
         type: 'Select',
         title: _t('editor.layers.analysis-form.weight-column'),
-        options: this._getColumns()
+        options: this._getColumnsByType('number')
       },
       category_column: {
         type: 'Select',
         title: _t('editor.layers.analysis-form.category-column'),
-        options: this._getColumns()
+        options: this._getColumnsByType()
       },
       aggregate: {
-        type: 'Select',
+        type: 'Enabler',
         title: _t('editor.layers.analysis-form.aggregate'),
-        options: ['COUNT', 'SUM', 'AVG', 'MIN', 'MAX']
+        validators: []
       }
-    });
-  },
+    };
 
-  _getColumns: function () {
-    var nodeDefModel = this._layerDefinitionModel.findAnalysisDefinitionNodeModel(this.get('source'));
-    var querySchemaModel = nodeDefModel.querySchemaModel;
-
-    var sourceColumns = querySchemaModel.columnsCollection.map(function (columnModel) {
-      var columnName = columnModel.get('name');
-
-      return {
-        val: columnName,
-        label: columnName,
-        type: columnModel.get('type')
-      };
-    });
-
-    return sourceColumns;
+    if (this.get('aggregate')) {
+      schema = _.extend(schema, {
+        aggregate_column: {
+          type: 'Select',
+          title: _t('editor.layers.analysis-form.aggregate'),
+          options: this._getColumnsByType('number')
+        },
+        aggregate_operation: {
+          type: 'Select',
+          title: _t('editor.layers.analysis-form.aggregate'),
+          options: ['count', 'sum', 'avg', 'min', 'max']
+        }
+      });
+    }
+    BaseAnalysisFormModel.prototype._setSchema.call(this, schema);
   }
 });
