+++ lib/assets/locale/en.json
@@ -497,37 +497,39 @@
         "hover": "Hover"
       },
       "analysis-form": {
+        "aggregate": "Aggregate",
+        "apply-btn": "Apply",
+        "boundaries": "Boundaries",
         "by": "By",
-        "by-walk": "walk",
         "by-bike": "bike",
         "by-car": "car",
-        "tracts": "Tracts",
-        "boundaries": "Boundaries",
-        "intact": "Intact",
+        "by-walk": "walk",
+        "category-column": "Category column",
+        "clusters-num": "# Clusters",
+        "create-btn": "Create",
         "dissolved": "Dissolved",
         "distance": "Distance",
         "input": "Input",
+        "input-layer": "Define your input layer",
+        "intact": "Intact",
         "kilometers": "km",
         "meters": "m",
         "miles": "mi",
+        "numerator": "Numerator",
         "parameters": "Parameters",
         "parameters-description": "Tune your analysis",
         "placeholder-text": "You have not added any analysis yet. Add new analysis to discover new things.",
-        "polygons_source": "Polygon source",
         "points_source": "Point source",
+        "polygons_source": "Polygon source",
         "radius": "Radius",
         "reference-layer-pluralize": "Define your reference layer |||| Define your reference layers",
-        "input-layer": "Define your input layer",
-        "tune-clusters": "Tune your clusters",
+        "significance": "Significance",
         "time": "Time",
+        "tracts": "Tracts",
+        "tune-clusters": "Tune your clusters",
         "type": "Type",
-        "workflow": "Workflow",
-        "apply-btn": "Apply",
-        "create-btn": "Create",
-        "clusters-num": "# Clusters",
-        "aggregate": "Aggregate",
         "weight-column": "Weight column",
-        "category-column": "Category column"
+        "workflow": "Workflow"
       },
       "infowindow": {
         "style": {
